requirements.txt
langchain
langchain-community
langchain-core
langchain-chroma
langchain-classic
python-telegram-bot>=20.0
python-dotenv
chromadb
sentence-transformers
langchain-groq
Flask
